+++ App/SplashScreenWindow.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows;
+
+namespace JocysCom.VS.AiCompanion
+{
+	/// <summary>
+	/// Interaction logic for SplashScreenWindow.xaml
+	/// </summary>
+	public partial class SplashScreenWindow : Window
+	{
+		public SplashScreenWindow()
+		{
+			InitializeComponent();
+		}
+	}
+}