+++ Documents/ZipFiles.Files.ps1
@@ -1,2 +1,2 @@
-.\ZipFiles.ps1 "$($PSScriptRoot)\Files" "$($PSScriptRoot)\Files\JocysCom.VS.AiCompanion.App.zip" JocysCom.VS.AiCompanion.App.exe
-.\ZipFiles.ps1 "$($PSScriptRoot)\Files" "$($PSScriptRoot)\Files\JocysCom.VS.AiCompanion.Extension.zip" JocysCom.VS.AiCompanion.Extension.vsix
+.\ZipFiles.ps1 "$($PSScriptRoot)\Files" "$($PSScriptRoot)\Files\JocysCom.VS.AiCompanion.App.zip" JocysCom.VS.AiCompanion.App.exe $true
+.\ZipFiles.ps1 "$($PSScriptRoot)\Files" "$($PSScriptRoot)\Files\JocysCom.VS.AiCompanion.Extension.zip" JocysCom.VS.AiCompanion.Extension.vsix $true